requirements.txt
dwave-ocean-sdk >= 6.8.0
matplotlib==3.7.4
